refactor: split syscall tests into their own files
